Split `NameMangler.DSL` up into submodules
