+++ hs-bindgen/hs-bindgen.cabal
@@ -109,6 +109,9 @@ library
       HsBindgen.GenTests.Internal
       HsBindgen.GenTests.Readme
       HsBindgen.Hs.NameMangler.API
+      HsBindgen.Hs.NameMangler.DSL.ReservedNames
+      HsBindgen.Hs.NameMangler.DSL.GenerateName
+      HsBindgen.Hs.NameMangler.DSL.Overrides
       HsBindgen.Hs.Translation
       HsBindgen.Imports
       HsBindgen.Orphans
